cuGraph/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==2.4.1
async-timeout==5.0.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.2.1
cuda-python==12.6.0
cudf-cu12==24.12.0
cugraph-cu12==24.12.0
cupy-cuda12x==13.3.0
cycler==0.12.1
dask==2024.11.2
dask-cuda==24.12.0
dask-cudf-cu12==24.12.0
dask-expr==1.1.19
debugpy==1.8.2
decorator==5.1.1
distributed==2024.11.2
distributed-ucxx-cu12==0.41.0
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastrlock==0.8.3
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
gensim==4.3.2
grip==4.6.2
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
libcudf-cu12==24.12.0
libkvikio-cu12==24.12.1
libucx-cu12==1.17.0.post1
libucxx-cu12==0.41.0
lightgbm==4.4.0
llvmlite==0.43.0
locket==1.0.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numba-cuda==0.0.17.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.19.3
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
nvtx==0.2.10
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
path-and-address==2.0.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
Pygments==2.18.0
pylibcudf-cu12==24.12.0
pylibcugraph-cu12==24.12.0
pylibraft-cu12==24.12.0
pynvjitlink-cu12==0.4.0
pynvml==11.4.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
raft-dask-cu12==24.12.0
rapids-dask-dependency==24.12.0
requests==2.32.3
rich==13.9.4
rmm-cu12==24.12.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.22.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12.1
tblib==3.0.0
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==1.0.0
torch @ http://snowfdc.deepln.com:8280/pytorch_whl_file/torch-2.2.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=1adf430f01ff649c848ac021785e18007b0714fdde68e4e65bd0c640bf3fb8e1
torchaudio @ http://snowfdc.deepln.com:8280/pytorch_whl_file/torchaudio-2.2.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=23f6236429e2bf676b820e8e7221a1d58aaf908bff2ba2665aa852df71a97961
torchvision @ http://snowfdc.deepln.com:8280/pytorch_whl_file/torchvision-0.17.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=27af47915f6e762c1d44e58e8088d22ac97445668f9f793524032b2baf4f34bd
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
ucx-py-cu12==0.41.0
ucxx-cu12==0.41.0
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
